requirements.txt
streamlit
pymupdf
pdf2image
pytesseract
Pillow
python-dotenv
google-generativeai
